add lst levels

+++ core/framework/levels.py
@@ -119,10 +119,11 @@ def write_start_info(level_path, message, file_name=None, file_content=None):
 
 def delete_start_files():
     '''Deletes the start files of a level. This function should be called upon level destruction.'''
-    try:
-        shutil.rmtree('start')
-    except Exception:
-        print("ERROR:Error deleting start files")
+    if os.path.exists('start'):
+        try:
+            shutil.rmtree('start')
+        except Exception:
+            print("ERROR:Error deleting start files")
 
 
 def generate_level_docs():
